Added resize in all directions for sticky and comment
